Crate cranelift_bitset

Bitsets for Cranelift.

This module provides two bitset implementations:

ScalarBitSet: A small bitset built on top of a single integer.

CompoundBitSet: A bitset that can store more bits than fit in a single integer, but which internally has heap allocations.

Re-exports

pub use compound::CompoundBitSet;

pub use scalar::ScalarBitSet;

Modules

compound

Compound bit sets.

scalar

Scalar bitsets.